move downsample_2d to downsamplers.py
