+++ src/diffusers/models/samplers/__init__.py
@@ -1,3 +1,3 @@
-from .downsamplers import Downsample1D, Downsample2D, FirDownsample2D, KDownsample2D
+from .downsamplers import Downsample1D, Downsample2D, FirDownsample2D, KDownsample2D, downsample_2d
 from .upsamplers import FirUpsample2D, KUpsample2D, Upsample1D, Upsample2D
-from .utils import downsample_2d, upfirdn2d_native, upsample_2d
+from .utils import upfirdn2d_native, upsample_2d
